% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumSR.R
\name{add_residuals}
\alias{add_residuals}
\title{Add residuals to \code{glmnet} object}
\usage{
add_residuals(mod, X, y, s = "lambda.min")
}
\description{
Add residuals to \code{glmnet} object
}
\keyword{internal}
