% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{get.cv.detectors}
\alias{get.cv.detectors}
\title{get mosum detectors}
\usage{
get.cv.detectors(
  X,
  y,
  G,
  lambda,
  family = c("gaussian", "binomial", "poisson"),
  grid.resolution = 1/G,
  ncores = 1,
  ...
)
}
\description{
get mosum detectors
}
\keyword{internal}
