% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{moseg.cv}
\alias{moseg.cv}
\title{Detect and estimate multiple change points in the sparse regression model,
selecting the number of change points using sample splitting}
\usage{
moseg.cv(
  X,
  y,
  G = NULL,
  lambda = NULL,
  max.cps = NULL,
  family = c("gaussian", "binomial", "poisson"),
  loss = c("1", "2"),
  folds = 2,
  path.length = 5,
  grid.resolution = 1/G,
  nu = 0.5,
  do.plot = TRUE,
  do.scale = TRUE,
  do.refinement = TRUE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{G}{integer bandwidth; default is chosen based on dimensions of \code{X}}

\item{lambda}{vector of numeric regularisation parameters}

\item{max.cps}{maximum number of change points to consider}

\item{family}{response type, one of \code{"gaussian","binomial","poisson"}}

\item{loss}{l-norm for CV loss function, one of \code{"1", "2"}}

\item{folds}{number of folds for CV}

\item{path.length}{number of \code{lambda} values to consider}

\item{grid.resolution}{controls number of subsamples to take}

\item{nu}{numeric localisation tuning parameter}

\item{do.plot}{Boolean - return plots}

\item{do.scale}{Boolean - centre and scale \code{X,y}}

\item{do.refinement}{Boolean - perform location refinement}

\item{ncores}{number of parallel cores}

\item{...}{optional arguments to \code{glmnet}}
}
\value{
List containing
\itemize{
  \item{\code{mosum}}{ numeric vector of mosum detector}
  \item{\code{cps}}{ integer vector of estimated change points}
  \item{\code{refined_cps}}{ integer vector of refined change points}
  \item{\code{lambda}}{ selected regularisation parameter}
  \item{\code{threshold}}{ implied threshold}
  \item{\code{detectors}}{ mosum detector series for each \code{lambda} value}
  \item{\code{cv}}{ matrix of cross-validation errors}
  \item{\code{model_list}}{ list of fitted piecewise models}
  \item{\code{plots}}{ list of detector and refinement plots}
  \item{\code{family}}{ input}
}
}
\description{
Detect and estimate multiple change points in the sparse regression model,
selecting the number of change points using sample splitting
}
\examples{
eqX <- eqdata[,-c(1,9)]
eq_thr <- moseg.cv(as.matrix(eqX), as.matrix(eqdata[,9]), G=120, max.cps = 3, ncores = 2)
}
