% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumSR.R
\name{moseg.fit}
\alias{moseg.fit}
\title{Fit a piecewise sparse regression model and evaluates a penalty. Fits a model to each stationary segment using \code{glmnet::cv.glmnet}.}
\usage{
moseg.fit(
  X,
  y,
  cps,
  lambda = c("min", "1se"),
  family = c("gaussian", "binomial", "poisson"),
  type = c("link", "response", "coefficients", "nonzero", "class"),
  do.plot = TRUE,
  do.scale = TRUE,
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{cps}{vector of change point locations}

\item{lambda}{regularisation parameter; either a numeric, or one of \code{"min","1se"} (see \link[glmnet]{cv.glmnet})}

\item{family}{response type, one of \code{"gaussian","binomial","poisson"}}

\item{type}{type of prediction; see \code{?glmnet.predict}}

\item{do.plot}{Boolean - return coefficent plot}

\item{do.scale}{Boolean - scale \code{X,y}}

\item{...}{optional arguments to \code{glmnet}}
}
\value{
list containing
\itemize{
  \item{\code{model}}{ list of fitted models}
  \item{\code{likelihood}}{ likelihood value}
  \item{\code{preds}}{ vector of fitted values for each time step}
  \item{\code{coeffs}}{ coefficient matrix}
  \item{\code{plot}}{ coefficient heatmap}
}
}
\description{
Fit a piecewise sparse regression model and evaluates a penalty. Fits a model to each stationary segment using \code{glmnet::cv.glmnet}.
}
\examples{
eqX <- eqdata[,-c(1,9)]
eq_mosum <- moseg.fit(as.matrix(eqX), as.matrix(eqdata[,9]), c(500))
}
