% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{moseg.ms.cv}
\alias{moseg.ms.cv}
\title{Detect and estimate multiple change points in the sparse regression model using multiple bandwidths,
selecting the number of change points using sample splitting}
\usage{
moseg.ms.cv(
  X,
  y,
  Gset = NULL,
  lambda = NULL,
  family = c("gaussian", "binomial", "poisson"),
  loss = c("1", "2"),
  folds = 1,
  threshold = NULL,
  grid.resolution = 1/Gset,
  nu = 0.5,
  do.plot = TRUE,
  do.scale = TRUE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{Gset}{integer vector of bandwidths; default is chosen based on dimensions of \code{X}}

\item{lambda}{regularisation parameter; either a numeric, or one of \code{"min","1se"} (see \link[glmnet]{cv.glmnet})}

\item{family}{response type, one of \code{"gaussian","binomial","poisson"}}

\item{loss}{l-norm for CV loss function, one of \code{"1", "2"}}

\item{folds}{number of folds for CV}

\item{threshold}{numeric test rejection threshold; see reference for default choice}

\item{grid.resolution}{controls number of subsamples to take}

\item{nu}{numeric localisation tuning parameter}

\item{do.plot}{Boolean - return plots}

\item{do.scale}{Boolean - scale \code{X,y}}

\item{ncores}{number of parallel cores}

\item{...}{optional arguments to \code{glmnet}}
}
\value{
\code{moseg.ms.cv} object containing
\itemize{
  \item{\code{anchors}}{ integer vector of estimated change points}
  \item{\code{refined.cps}}{ integer vector of refined change points}
  \item{\code{moseg.G}}{ list of `moseg.cv` objects corresponding to `Gset` in ascending order}
}
}
\description{
Detect and estimate multiple change points in the sparse regression model using multiple bandwidths,
selecting the number of change points using sample splitting
}
\examples{
eqX <- eqdata[,-c(1,9)]
eq_mosum <- moseg.ms.cv(as.matrix(eqX), eqdata[,9], c(60,90,120), ncores = 2)
}
