% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA.R
\name{moseg.ms}
\alias{moseg.ms}
\title{Detect and estimate multiple change points in the sparse regression model using multiple bandwidths}
\usage{
moseg.ms(
  X,
  y,
  Gset,
  lambda = c("min", "1se"),
  family = c("gaussian", "binomial", "poisson"),
  threshold = NULL,
  grid.resolution = NULL,
  nu = 0.5,
  do.plot = TRUE,
  do.scale = TRUE,
  ncores = 1,
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{Gset}{integer vector of bandwidths; default is chosen based on dimensions of \code{X}}

\item{lambda}{regularisation parameter; either a numeric, or one of \code{"min","1se"} (see \link[glmnet]{cv.glmnet})}

\item{family}{response type, one of \code{"gaussian","binomial","poisson"}}

\item{threshold}{numeric test rejection threshold; see  \link[moseg]{moseg} for default choice}

\item{grid.resolution}{controls number of subsamples to take}

\item{nu}{numeric localisation tuning parameter}

\item{do.plot}{Boolean - return plots}

\item{do.scale}{Boolean - scale \code{X,y}}

\item{ncores}{number of parallel cores}

\item{...}{optional arguments to \code{glmnet}}
}
\value{
List containing
\itemize{
  \item{\code{cps}}{ integer vector of estimated change points}
  \item{\code{plot}}{ multiscale plot}
  \item{\code{moseg.G}}{ list of `moseg` objects corresponding to `Gset` in ascending order}
}
}
\description{
Detect and estimate multiple change points in the sparse regression model using multiple bandwidths
}
\examples{
eqX <- eqdata[,-c(1,9)]
eq_mosum <- moseg.ms(as.matrix(eqX), eqdata[,9], c(60,90,120), ncores = 2)
}
\seealso{
\link[moseg]{moseg.ms.cv}
}
