% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumSR.R
\name{moseg}
\alias{moseg}
\title{Detect and estimate multiple change points in the sparse regression model}
\usage{
moseg(
  X,
  y,
  G = NULL,
  lambda = c("min", "1se"),
  family = c("gaussian", "binomial", "poisson"),
  threshold = NULL,
  n.cps = NULL,
  grid.resolution = 1/G,
  nu = 0.5,
  do.refinement = TRUE,
  do.plot = TRUE,
  do.scale = TRUE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{G}{integer bandwidth; default is chosen based on dimensions of \code{X}}

\item{lambda}{regularisation parameter; either a numeric, or one of \code{"min","1se"} (see \link[glmnet]{cv.glmnet})}

\item{family}{response type, one of \code{"gaussian","binomial","poisson"}}

\item{threshold}{numeric test rejection threshold; see details for default}

\item{n.cps}{chosen number of change points to return; if specified, overrides \code{threshold}}

\item{grid.resolution}{controls number of subsamples to take}

\item{nu}{numeric localisation tuning parameter}

\item{do.refinement}{Boolean - perform location refinement}

\item{do.plot}{Boolean - return plots}

\item{do.scale}{Boolean - centre and scale \code{X,y}}

\item{ncores}{number of parallel cores}

\item{...}{optional arguments to \code{glmnet}}
}
\value{
List containing
\itemize{
  \item{\code{mosum}}{ numeric vector of mosum detector}
  \item{\code{cps}}{ integer vector of estimated change points}
  \item{\code{refined.cps}}{ integer vector of refined change points}
  \item{\code{G}}{ input}
  \item{\code{lambda}}{ input}
  \item{\code{threshold}}{ input}
  \item{\code{grid.resolution}}{ input}
  \item{\code{family}}{ input}
  \item{\code{plots}}{ list of detector and refinement plots}
  \item{\code{model_list}}{ list of fitted \code{glmnet} models at each grid point}
}
}
\description{
Detect and estimate multiple change points in the sparse regression model
}
\details{
The default threshold is chosen as a product of exponents of \code{lambda} and \code{G}.
 For a more accurate, but slightly slower, procedure, see \link[moseg]{moseg.cv}.
}
\examples{
eqX <- eqdata[,-c(1,9)]
eq_mosum <- moseg(as.matrix(eqX), eqdata[,9], 120, grid.resolution = 1/10, ncores = 2)
}
