% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{moseg.sim}
\alias{moseg.sim}
\title{Simulate from a sparse piecewise regression model}
\usage{
moseg.sim(
  n,
  p,
  sparsity = floor(sqrt(p)),
  q = 1,
  sigma.noise = 1,
  sigma.x = c("id", "band", "ar"),
  kappa = 1
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of parameters}

\item{sparsity}{number of non-zero parameters}

\item{q}{number of change points}

\item{sigma.noise}{error standard deviation}

\item{sigma.x}{covariance structure of \code{X}, one of \code{"id", "band", "ar"}}

\item{kappa}{change size}
}
\value{
List containing
\itemize{
  \item{\code{y}}{ response vector}
  \item{\code{X}}{ matrix of covariates}
  \item{\code{cps}}{ vector of change points}
  \item{\code{beta}}{ matrix of parameters corresponding to each regime}
}
and all inputs
}
\description{
Simulate from a sparse piecewise regression model
}
\examples{
moseg.sim(100, 50)
}
