% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA.R
\name{mosumsr.multiscale}
\alias{mosumsr.multiscale}
\title{Detect and estimate multiple change points in the sparse regression model using multiple bandwidths}
\usage{
mosumsr.multiscale(
  X,
  y,
  Gset,
  lambda = c("min", "1se"),
  family = c("gaussian", "binomial", "poisson"),
  threshold = NULL,
  grid.resolution = NULL,
  nu = 0.5,
  do.plot = TRUE,
  do.scale = TRUE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{Gset}{integer vector of bandwidths; default smallest is \code{round(30 + ncol(X)/100)}}

\item{lambda}{regularisation parameter; either a numeric, or one of \code{"min","1se"} (see \link[glmnet]{cv.glmnet})}

\item{family}{response type, one of \code{"gaussian","binomial","poisson"}}

\item{threshold}{numeric test rejection threshold; see  \link[mosumsr]{mosumsr} for default choice}

\item{grid.resolution}{controls number of subsamples to take}

\item{nu}{numeric localisation tuning parameter}

\item{do.plot}{Boolean - return plots}

\item{do.scale}{Boolean - scale \code{X,y}}

\item{ncores}{number of parallel cores}

\item{...}{optional arguments to \code{glmnet}}
}
\value{
List containing
\itemize{
  \item{\code{cps}}{ integer vector of estimated change points}
  \item{\code{plot}}{ multiscale plot}
  \item{\code{mosumsr.G}}{ list of `mosumsr` objects corresponding to `Gset` in ascending order}
}
}
\description{
Detect and estimate multiple change points in the sparse regression model using multiple bandwidths
}
\examples{
eqX <- eqdata[,-c(1,9)]
eq_mosum <- mosumsr.multiscale(as.matrix(eqX), eqdata[,9], c(60,90,120), ncores = 2)
}
\seealso{
\link[mosumsr]{mosumsr.multiscale.cv}
}
