% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{plot.moseg.cv}
\alias{plot.moseg.cv}
\title{Plot the moseg detector}
\usage{
\method{plot}{moseg.cv}(x, type = c("cv", "mosum"), ...)
}
\arguments{
\item{x}{\code{moseg.cv} object}

\item{type}{plot type, one of \code{"cv","mosum"}}

\item{...}{unused}
}
\value{
A cv or detector plot
}
\description{
Plotting method for S3 objects of class \code{moseg.cv}.
}
