% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{plot.moseg.ms.cv}
\alias{plot.moseg.ms.cv}
\title{Plot the multiscale moseg.cv detector}
\usage{
\method{plot}{moseg.ms.cv}(x, type = c("cv", "mosum"), ...)
}
\arguments{
\item{x}{\code{moseg.ms} object}

\item{type}{plot type, one of \code{"cv","mosum"}}

\item{...}{unused}
}
\value{
cv or detector plots
}
\description{
Plotting method for S3 objects of class \code{moseg.ms.cv}.
}
