% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA.R
\name{plot.moseg.ms}
\alias{plot.moseg.ms}
\title{Plot the multiscale moseg detector}
\usage{
\method{plot}{moseg.ms}(x, ...)
}
\arguments{
\item{x}{\code{moseg.ms} object}

\item{...}{unused}
}
\value{
Detector plots
}
\description{
Plotting method for S3 objects of class \code{moseg.ms}.
}
