% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumSR.R
\name{plot.moseg}
\alias{plot.moseg}
\title{Plot the moseg detector}
\usage{
\method{plot}{moseg}(x, ...)
}
\arguments{
\item{x}{\code{moseg} object}

\item{...}{unused}
}
\value{
A detector plot
}
\description{
Plotting method for S3 objects of class \code{moseg}.
}
