% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostep.R
\name{refinement}
\alias{refinement}
\title{Refine a change point location estimate}
\usage{
refinement(
  X,
  y,
  k,
  G,
  lambda,
  l_mod,
  r_mod,
  L_min = NULL,
  U_max = NULL,
  family = c("gaussian", "binomial", "poisson")
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{k}{integer change point location estimate}

\item{G}{integer bandwidth}

\item{lambda}{numeric regularisation parameter}

\item{l_mod}{\code{glmnet} object from left of change point}

\item{r_mod}{\code{glmnet} object from right of change point}

\item{family}{response type, one of \code{"gaussian","binomial","poisson"}}
}
\value{
list containing
 \itemize{
   \item{\code{cp}}{ refined change point estimate}
   \item{\code{objective}}{ Loss function}
}
}
\description{
Refine a change point location estimate
}
\keyword{internal}
