% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fit.cv}
\alias{fit.cv}
\title{fit models with cross-validation}
\usage{
fit.cv(
  X,
  y,
  cps,
  ranks,
  max.cps,
  lambda,
  family = c("gaussian", "binomial", "poisson"),
  do.plot = TRUE,
  do.scale = TRUE,
  ...
)
}
\description{
fit models with cross-validation
}
\keyword{internal}
